x86, PCI, ACPI: Add proper type casts to printk arguments
On 32bit the resource_size_t is u32 so gcc complains stupidly:

arch/x86/pci/acpi.c:271:4: warning: format ‘%llx’ expects argument of type ‘long long unsigned
int’, but argument 3 has type ‘resource_size_t’ [-Wformat=]

Fixes: commit e22ce93870de ("x86, PCI, ACPI: Kill private function resource_to_addr() in arch/x86/pci/acpi.c"
Reported-by: Borislav Petkov <bp@alien8.de>
Cc: Jiang Liu <jiang.liu@linux.intel.com>

+++ arch/x86/pci/acpi.c
@@ -270,12 +270,13 @@ static acpi_status setup_resource(struct acpi_resource *acpi_res, void *data)
 		if (r.end <= r.start) {
 			dev_info(&info->bridge->dev,
 				"host bridge window [%#llx-%#llx] (ignored, not CPU addressable)\n",
-				 r.start, orig_end);
+				 (unsigned long long)r.start, orig_end);
 			return AE_OK;
 		} else if (orig_end != r.end) {
 			dev_info(&info->bridge->dev,
 				"host bridge window [%#llx-%#llx] ([%#llx-%#llx] ignored, not CPU addressable)\n",
-				r.start, orig_end, r.end + 1, orig_end);
+				 (unsigned long long)r.start, orig_end,
+				 (unsigned long long)r.end + 1, orig_end);
 		}
 	}
